viagens/requirements.txt
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.3
backcall==0.2.0
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
cssutils==1.0.2
decorator==4.4.2
email-to==0.1.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-spreadsheet==0.0.6
googleapis-common-protos==1.52.0
httplib2==0.18.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
itemadapter==0.1.0
jedi==0.17.2
lxml==4.5.2
Markdown==3.3.3
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.5
parsel==1.6.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
premailer==3.7.0
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
queuelib==1.5.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
Scrapy==2.2.1
scrapy-splash==0.7.2
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
toml==0.10.1
traitlets==4.3.3
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
zipp==3.4.0
zope.interface==5.1.0
